requirements.txt
huggingface_hub
datasets
pandas
matplotlib
seaborn
torch
accelerate
tqdm
transformers>=4.22.0
torchvision
evaluate
scikit-learn==1.5.2
wandb
timm
python-dotenv
annoy 
renumics-spotlight
cleanlab
cleanvision
onnxruntime==1.15.1